Separate kanjis/words from kanas
The first tab of the App is reserved to learning the
scripts and associated meanings. Let's introduce
a ScriptMode enum rather than a string, with 3
variants: Hiragana, Katakana, Kanji. The new variant, Kanji, will have a list of single kanji words
(~300 most common). This is more suitable for a
beginner level compared to the first most common
words with hard kanjis to understand/remember.