Add a grid selection view for kanas
Currently, only one way recognition is supported, which is from kana/kanji to transliteration/translation.

This commit adds a way to improve recognition, by letting the user select the right kana/kanji between a set of 6, based on the provided transliteration/translation.

The set of 6 can be choosen in different way:
- Randomly
- From the top kana/kanjis most similar
- From the previously confused kanas

The third way, is very personalised and thus preferrable. The app already tracks statistics on mistakes, so we can add a set that tracks which kanas are routenily confused with each other. We can mix this with random kanas too.